support pam2xfl and xfl2pam

+++ PopStudio.Avalonia/Pages/Control_Pam.axaml.cs
@@ -5,7 +5,6 @@ using Avalonia.Threading;
 using PopStudio.Language.Languages;
 using System.Diagnostics;
 using PopStudio.Platform;
-using System.Xml.Linq;
 
 namespace PopStudio.Avalonia.Pages
 {
@@ -16,6 +15,20 @@ namespace PopStudio.Avalonia.Pages
             InitializeComponent();
             LoadControl();
             LoadFont();
+            CB_InMode.Items = new List<string>
+            {
+                "Raw_Binary",
+                "Studio_Json",
+                "Flash_Xfl"
+            };
+            CB_InMode.SelectedIndex = 0;
+            CB_OutMode.Items = new List<string>
+            {
+                "Raw_Binary",
+                "Studio_Json",
+                "Flash_Xfl"
+            };
+            CB_OutMode.SelectedIndex = 1;
             MAUIStr.OnLanguageChanged += LoadFont;
         }
 
@@ -35,11 +48,10 @@ namespace PopStudio.Avalonia.Pages
             label_batch2 = this.Get<TextBlock>("label_batch2");
             batch_mode = this.Get<ToggleSwitch>("batch_mode");
             label_introduction = this.Get<TextBlock>("label_introduction");
-            label_choosemode = this.Get<TextBlock>("label_choosemode");
-            label_mode1 = this.Get<TextBlock>("label_mode1");
-            label_mode2 = this.Get<TextBlock>("label_mode2");
             text1 = this.Get<TextBlock>("text1");
             text2 = this.Get<TextBlock>("text2");
+            text_in = this.Get<TextBlock>("text_in");
+            text_out = this.Get<TextBlock>("text_out");
             button1 = this.Get<Button>("button1");
             button2 = this.Get<Button>("button2");
             button_run = this.Get<Button>("button_run");
@@ -47,7 +59,8 @@ namespace PopStudio.Avalonia.Pages
             text4 = this.Get<TextBlock>("text4");
             textbox1 = this.Get<TextBox>("textbox1");
             textbox2 = this.Get<TextBox>("textbox2");
-            TB_Mode = this.Get<ToggleSwitch>("TB_Mode");
+            CB_InMode = this.Get<ComboBox>("CB_InMode");
+            CB_OutMode = this.Get<ComboBox>("CB_OutMode");
         }
 
         void LoadFont()
@@ -55,62 +68,23 @@ namespace PopStudio.Avalonia.Pages
             label_batch1.Text = MAUIStr.Obj.Share_SingleMode;
             label_batch2.Text = MAUIStr.Obj.Share_BatchMode;
             label_introduction.Text = MAUIStr.Obj.Pam_Introduction;
-            label_choosemode.Text = MAUIStr.Obj.Share_ChooseMode;
             if (batch_mode.IsChecked == true)
             {
-                label_mode1.Text = MAUIStr.Obj.Pam_Mode1_Batch;
-                label_mode2.Text = MAUIStr.Obj.Pam_Mode2_Batch;
+                text1.Text = MAUIStr.Obj.Pam_Choose1_Batch;
+                text2.Text = MAUIStr.Obj.Pam_Choose2_Batch;
             }
             else
             {
-                label_mode1.Text = MAUIStr.Obj.Pam_Mode1;
-                label_mode2.Text = MAUIStr.Obj.Pam_Mode2;
+                text1.Text = MAUIStr.Obj.Pam_Choose1;
+                text2.Text = MAUIStr.Obj.Pam_Choose2;
             }
-            LoadFont_Checked(TB_Mode.IsChecked == true);
+            text_in.Text = MAUIStr.Obj.Pam_InFormat;
+            text_out.Text = MAUIStr.Obj.Pam_OutFormat;
             button1.Content = MAUIStr.Obj.Share_Choose;
             button2.Content = MAUIStr.Obj.Share_Choose;
             button_run.Content = MAUIStr.Obj.Share_Run;
             label_statue.Text = MAUIStr.Obj.Share_RunStatue;
             text4.Text = MAUIStr.Obj.Share_Waiting;
-        }
-
-        void LoadFont_Checked(bool v)
-        {
-            if (batch_mode.IsChecked == true)
-            {
-                if (v)
-                {
-                    text1.Text = MAUIStr.Obj.Pam_Choose4_Batch;
-                    text2.Text = MAUIStr.Obj.Pam_Choose5_Batch;
-                }
-                else
-                {
-                    text1.Text = MAUIStr.Obj.Pam_Choose1_Batch;
-                    text2.Text = MAUIStr.Obj.Pam_Choose2_Batch;
-                }
-            }
-            else
-            {
-                if (v)
-                {
-                    text1.Text = MAUIStr.Obj.Pam_Choose4;
-                    text2.Text = MAUIStr.Obj.Pam_Choose5;
-                }
-                else
-                {
-                    text1.Text = MAUIStr.Obj.Pam_Choose1;
-                    text2.Text = MAUIStr.Obj.Pam_Choose2;
-                }
-            }
-        }
-
-        private void Switch_Checked(object sender, RoutedEventArgs e)
-        {
-            if (sender is ToggleSwitch s)
-            {
-                LoadFont_Checked(s.IsChecked == true);
-                (textbox1.Text, textbox2.Text) = (textbox2.Text, textbox1.Text);
-            }
         }
 
         private void Switch_Batch_Checked(object sender, RoutedEventArgs e)
@@ -122,7 +96,7 @@ namespace PopStudio.Avalonia.Pages
         {
             try
             {
-                string val = batch_mode.IsChecked == false ? (await new OpenFileDialog { AllowMultiple = false }.ShowAsync(MainWindow.Singleten))?[0] : (await new OpenFolderDialog().ShowAsync(MainWindow.Singleten));
+                string val = (CB_InMode.SelectedIndex == 2 || batch_mode.IsChecked == true) ? (await new OpenFolderDialog().ShowAsync(MainWindow.Singleten)) : (await new OpenFileDialog { AllowMultiple = false }.ShowAsync(MainWindow.Singleten))?[0];
                 if (!string.IsNullOrEmpty(val)) textbox1.Text = val;
             }
             catch (Exception)
@@ -134,7 +108,7 @@ namespace PopStudio.Avalonia.Pages
         {
             try
             {
-                string val = batch_mode.IsChecked == false ? (await new SaveFileDialog().ShowAsync(MainWindow.Singleten)) : (await new OpenFolderDialog().ShowAsync(MainWindow.Singleten));
+                string val = (CB_OutMode.SelectedIndex == 2 || batch_mode.IsChecked == true) ? await new OpenFolderDialog().ShowAsync(MainWindow.Singleten) : await new SaveFileDialog().ShowAsync(MainWindow.Singleten);
                 if (!string.IsNullOrEmpty(val)) textbox2.Text = val;
             }
             catch (Exception)
@@ -147,10 +121,10 @@ namespace PopStudio.Avalonia.Pages
             Button b = (Button)sender;
             b.IsEnabled = false;
             text4.Text = MAUIStr.Obj.Share_Running;
-            bool mode = TB_Mode.IsChecked == true;
             string inFile = textbox1.Text;
             string outFile = textbox2.Text;
-            //int cmode = CB_CMode.SelectedIndex;
+            int inmode = CB_InMode.SelectedIndex;
+            int outmode = CB_OutMode.SelectedIndex;
             bool batchmode = batch_mode.IsChecked == true;
             new Thread(new ThreadStart(() =>
             {
@@ -159,7 +133,13 @@ namespace PopStudio.Avalonia.Pages
                 sw.Start();
                 try
                 {
-                    string outFormat = mode ? ".pam" : ".pam.json";
+                    string outFormat = outmode switch
+                    {
+                        0 => ".pam",
+                        1 => ".pam.json",
+                        2 => ".xfl",
+                        _ => null
+                    };
                     if (batchmode)
                     {
                         if (!Directory.Exists(inFile))
@@ -170,7 +150,13 @@ namespace PopStudio.Avalonia.Pages
                         int length = inFile.Length;
                         string[] files = YFAPI.GetFiles(inFile);
                         YFAPI.NewDir(outFile);
-                        string rightFormat = mode ? ".pam.json" : ".pam";
+                        string rightFormat = inmode switch
+                        {
+                            0 => ".pam",
+                            1 => ".pam.json",
+                            2 => ".xfl",
+                            _ => null
+                        };
                         int rightFormatLength = rightFormat.Length;
                         foreach (string mfile in files)
                         {
@@ -182,41 +168,30 @@ namespace PopStudio.Avalonia.Pages
                             YFAPI.NewDir(newPath, false);
                             try
                             {
-                                if (mode)
-                                {
-                                    YFAPI.EncodePam(mfile, newPath);
-                                }
-                                else
-                                {
-                                    YFAPI.DecodePam(mfile, newPath);
-                                }
+                                YFAPI.Pam(mfile, newPath, inmode, outmode);
                             }
                             catch (Exception)
                             {
-                                File.Delete(newPath);
+                                if (outmode != 2)
+                                {
+                                    File.Delete(newPath);
+                                }
                             }
                         }
                     }
                     else
                     {
-                        if (!File.Exists(inFile))
+                        if (inmode != 2 && !File.Exists(inFile))
                         {
                             throw new Exception(string.Format(MAUIStr.Obj.Share_FileNotFound, inFile));
                         }
-                        if (Directory.Exists(outFile))
+                        if (outmode != 2 && Directory.Exists(outFile))
                         {
                             outFile += "/" + Path.GetFileName(inFile) + outFormat;
                             outFile = YFAPI.FormatPath(outFile);
                         }
-                        YFAPI.NewDir(outFile, false);
-                        if (mode)
-                        {
-                            YFAPI.EncodePam(inFile, outFile);
-                        }
-                        else
-                        {
-                            YFAPI.DecodePam(inFile, outFile);
-                        }
+                        YFAPI.NewDir(outFile, outmode == 2);
+                        YFAPI.Pam(inFile, outFile, inmode, outmode);
                     }
                 }
                 catch (Exception ex)
@@ -229,7 +204,7 @@ namespace PopStudio.Avalonia.Pages
                 {
                     if (err == null)
                     {
-                        text4.Text = string.Format(MAUIStr.Obj.Share_Finish, time);
+                        text4.Text = string.Format(MAUIStr.Obj.Share_Finish, time.ToString("F3"));
                     }
                     else
                     {
